hive: add option to persistently disallow substitutes
While Colmena provides a "--no-substitutes" flag to disable substitutes,
there is no way to persistently disable their use. Such an option would
be useful though in cases where one is deploying to a node that does not
have an internet connection, which would otherwise fail if the flag was
not given.

Introduce a new per-node "deployment.noSubstitute" option that allows
you to persistently disable the use of substitutes. This option defaults
to "false", which is in line with the current default. Furthermore, the
option gets overridden by the "--no-substitutes" command line flag.